requirements.txt
# Used by pip to install required python packages
# Usage: pip install -r requirements.txt

alembic~=1.6.5
certifi~=2021.5.30
charset-normalizer~=2.0.4
click~=8.0.1
Flask~=2.0.1
Flask-Migrate~=3.1.0
Flask-SQLAlchemy~=2.5.1
greenlet~=1.1.0
gunicorn
idna~=3.2
importlib-metadata~=1.5.0
itsdangerous~=2.0.1
Jinja2~=3.0.1
jinja2-time~=0.2.0
Mako~=1.1.4
MarkupSafe~=2.0.1
python-dateutil~=2.8.2
python-editor~=1.0.4
PyYAML~=5.4.1
requests~=2.26.0
six~=1.16.0
SQLAlchemy~=1.4.22
urllib3~=1.26.6
Werkzeug~=2.0.1
